--- inc/array2defs.h
@@ -1,6 +1,0 @@
-#ifndef ARRAY2DEFS_H
-#define ARRAY2DEFS_H 1
-#include "lispemul.h" /* for LispPTR */
-LispPTR N_OP_misc4(register LispPTR data, register LispPTR base, register LispPTR typenumber,
-                   register LispPTR inx, int alpha);
-#endif
