--- inc/array3defs.h
@@ -1,5 +1,0 @@
-#ifndef ARRAY3DEFS_H
-#define ARRAY3DEFS_H 1
-#include "lispemul.h" /* for LispPTR */
-LispPTR N_OP_aref1(register LispPTR arrayarg, register LispPTR inx);
-#endif
