--- inc/array4defs.h
@@ -1,5 +1,0 @@
-#ifndef ARRAY4DEFS_H
-#define ARRAY4DEFS_H 1
-#include "lispemul.h" /* for LispPTR */
-LispPTR N_OP_aset1(register LispPTR data, LispPTR arrayarg, register int inx);
-#endif
