--- inc/array5defs.h
@@ -1,5 +1,0 @@
-#ifndef ARRAY5DEFS_H
-#define ARRAY5DEFS_H 1
-#include "lispemul.h" /* for LispPTR */
-LispPTR N_OP_aref2(LispPTR arrayarg, LispPTR inx0, LispPTR inx1);
-#endif
