--- inc/array6defs.h
@@ -1,5 +1,0 @@
-#ifndef ARRAY6DEFS_H
-#define ARRAY6DEFS_H 1
-#include "lispemul.h" /* for LispPTR */
-LispPTR N_OP_aset2(register LispPTR data, LispPTR arrayarg, LispPTR inx0, LispPTR inx1);
-#endif
