--- inc/arraydefs.h
@@ -1,5 +1,0 @@
-#ifndef ARRAYDEFS_H
-#define ARRAYDEFS_H 1
-#include "lispemul.h" /* for LispPTR */
-LispPTR N_OP_misc3(LispPTR baseL, LispPTR typenumber, LispPTR inx, int alpha);
-#endif
