+++ src/xc.c
@@ -60,12 +60,7 @@
 #include "arith2defs.h"
 #include "arith3defs.h"
 #include "arith4defs.h"
-#include "arraydefs.h"
-#include "array2defs.h"
-#include "array3defs.h"
-#include "array4defs.h"
-#include "array5defs.h"
-#include "array6defs.h"
+#include "arrayopsdefs.h"
 #include "bitbltdefs.h"
 #include "bltdefs.h"
 #include "byteswapdefs.h"
